Enhance card and review functionality: add average rating calculation, update review validation, and implement Review seeder

+++ database/migrations/2025_01_07_081914_create_reviews_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained('users')->cascadeOnDelete();
             $table->foreignId('card_id')->constrained()->cascadeOnDelete();
-            $table->integer('rating');
+            $table->decimal('rating',2,1);
             $table->string('comment');
             $table->timestamps();
         });
